% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dsc2ace}
\alias{dsc2ace}
\title{convert DSC configuration to ACE editor content string}
\usage{
dsc2ace(dsc_list)
}
\arguments{
\item{dsc_list}{dsc output list}
}
\description{
convert DSC configuration to ACE editor content string
}

