% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_execs}
\alias{get_execs}
\title{get executable names for blocks}
\usage{
get_execs(dsc_list)
}
\arguments{
\item{dsc_list}{dsc output list}
}
\value{
a list containing block names and their executable names
}
\description{
get executable names for blocks
}

