% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_params}
\alias{get_params}
\title{get parameter names for executables}
\usage{
get_params(dsc_list)
}
\arguments{
\item{dsc_list}{dsc output list}
}
\value{
a list containing block names, executable names, and parameter names
}
\description{
get parameter names for executables
}

