% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinydsc.R
\name{shinydsc}
\alias{shinydsc}
\title{Run shinydsc app}
\usage{
shinydsc(display.mode = "normal")
}
\arguments{
\item{display.mode}{shiny app display mode parameter}
}
\description{
Run shinydsc app
}
\examples{
\donttest{shinydsc()}

}

